feat: migrate RichLexicalPostComposer functionality to PostComposer component for improved structure and maintainability
